refactor: split mount in parser
